+++ Database/DatabaseConfig.cs
@@ -1,5 +1,5 @@
 namespace LabManager.Database;
 
 class DatabaseConfig{
-    public string connectionString { get => "Data souce=database.db"; }
+    public string ConnectionString { get => "Data source=database.db"; }
 }
