+++ Program.cs
@@ -2,13 +2,7 @@
 using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
-
-Console.WriteLine(args);
-
-foreach (var arg in args)
-{
-    Console.WriteLine(arg);
-}
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 new DatabaseSetup(databaseConfig);
@@ -30,25 +24,19 @@ if(modelName =="Computer"){
         }
     }
     if(modelAction =="New"){
-        var conection = new SqliteConnection("Data Source=database.db");
-   
-        conection.Open();
-        
+      
         Console.WriteLine("New computer");
         int id = Convert.ToInt32(args[2]);
         string ram = args [3];
         string processor = args [4];
 
 
-    var command = conection.CreateCommand();
-    command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
-    command.Parameters.AddWithValue( "$id", id);
-    command.Parameters.AddWithValue( "$ram", ram);
-    command.Parameters.AddWithValue( "$processor", processor);
+        var computer = new Computer(id, ram, processor);
+        computerRepository.Save(computer);
 
-    command.ExecuteNonQuery();
-    conection.Close();
-        Console.WriteLine("{0}, {1}, {2}", id, ram, processor);
+       
+
+
 
     }
 }
